refactor: style 파일명 컨벤션으로 수정
