+++ frontend/src/components/Modal/style.tsx
