+++ frontend/src/components/Stepper/style.tsx
