+++ frontend/src/components/Template/index.tsx
@@ -1,5 +1,5 @@
 import { PropsWithChildren } from 'react';
-import { BaseTemplate, PageContainer } from './Template.style';
+import { BaseTemplate, PageContainer } from './style';
 import PageSideBar from '../SideBar/PageSideBar';
 
 const Template = ({ children }: PropsWithChildren) => {
